PlayerDropItemListener.java implemented
